# backend/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/auth/register": {
			"post": {
				"operationId": "AuthController_register",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"summary": "Đăng ký tài khoản mới (mặc định vai trò học viên)",
				"tags": ["Auth"]
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "AuthController_login",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"summary": "Đăng nhập và nhận cặp access/refresh token",
				"tags": ["Auth"]
			}
		},
		"/api/auth/refresh": {
			"post": {
				"operationId": "AuthController_refresh",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"summary": "Làm mới access token bằng refresh token (rotation)",
				"tags": ["Auth"]
			}
		},
		"/api/auth/logout": {
			"post": {
				"operationId": "AuthController_logout",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogoutDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Thu hồi refresh token hiện tại (hoặc mọi phiên)",
				"tags": ["Auth"]
			}
		},
		"/api/auth/forgot-password": {
			"post": {
				"operationId": "AuthController_forgotPassword",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"summary": "Yêu cầu đặt lại mật khẩu (luôn trả cùng kết quả)",
				"tags": ["Auth"]
			}
		},
		"/api/auth/reset-password": {
			"post": {
				"operationId": "AuthController_resetPassword",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"summary": "Đặt lại mật khẩu bằng token một lần",
				"tags": ["Auth"]
			}
		},
		"/api/auth/change-password": {
			"post": {
				"operationId": "AuthController_changePassword",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Đổi mật khẩu của người dùng đang đăng nhập",
				"tags": ["Auth"]
			}
		},
		"/api/auth/me": {
			"get": {
				"operationId": "AuthController_me",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Hồ sơ người dùng đang đăng nhập",
				"tags": ["Auth"]
			}
		},
...
		"/api/courses/{id}": {
			"get": {
				"operationId": "CourseController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Xem chi tiết khoá học kèm giảng viên và điều kiện tiên quyết",
				"tags": ["Courses"]
			},
			"patch": {
				"operationId": "CourseController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCourseDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Cập nhật khoá học (giảng viên phụ trách hoặc admin)",
				"tags": ["Courses"]
			},
			"delete": {
				"operationId": "CourseController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Xoá khoá học chưa có học viên ghi danh",
				"tags": ["Courses"]
			}
		},
...
		"/api/courses/{id}/cohorts": {
			"get": {
				"operationId": "CourseController_findCohorts",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Danh sách lớp của khoá học kèm sĩ số",
				"tags": ["Courses"]
			},
			"post": {
				"operationId": "CourseController_createCohort",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCohortDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Tạo lớp cho khoá học (giảng viên phụ trách hoặc admin)",
				"tags": ["Courses"]
			}
		},
...
		"/api/cohorts/{id}": {
			"patch": {
				"operationId": "CohortController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCohortDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Cập nhật lớp của khoá học mình phụ trách",
				"tags": ["Cohorts"]
			},
			"delete": {
				"operationId": "CohortController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"summary": "Xoá lớp (giữ nguyên ghi danh và phân công giảng viên)",
				"tags": ["Cohorts"]
			}
		},
...
	},
	"info": {
		"title": "UniPrep API Docs",
		"description": "UniPrep API documentation",
		"version": "1.0",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"RegisterDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"example": "sv2026001@hcmut.edu.vn",
						"description": "Email đăng nhập, không phân biệt hoa/thường",
						"maxLength": 255
					},
					"password": {
						"type": "string",
						"example": "Abcd@1234",
						"description": "Chính sách mật khẩu ở §5.1 (8–16 ký tự, hoa, thường, số, ký tự đặc biệt)"
					},
					"fullName": {
						"type": "string",
						"example": "Nguyễn Văn A",
						"minLength": 1,
						"maxLength": 100
					},
					"studentCode": {
						"type": "string",
						"example": "SV2026001",
						"description": "Chỉ dùng khi tạo học viên; duy nhất nếu có",
						"maxLength": 50
					}
				},
				"required": ["email", "password", "fullName"]
			},
			"ApiResponseDto": {
				"type": "object",
				"properties": {
					"error": {
						"type": "boolean",
						"example": false
					},
					"data": {
						"type": "object",
						"description": "Payload data (object, array, primitive) or null",
						"nullable": true
					},
					"message": {
						"type": "string",
						"example": "Success"
					}
				},
				"required": ["error", "message"]
			},
			"LoginDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"example": "sv2026001@hcmut.edu.vn"
					},
					"password": {
						"type": "string",
						"example": "Abcd@1234",
						"minLength": 1
					}
				},
				"required": ["email", "password"]
			},
			"RefreshTokenDto": {
				"type": "object",
				"properties": {
					"refreshToken": {
						"type": "string",
						"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
						"description": "Refresh token nhận được từ lần đăng nhập/refresh trước đó"
					}
				},
				"required": ["refreshToken"]
			},
			"LogoutDto": {
				"type": "object",
				"properties": {
					"refreshToken": {
						"type": "string",
						"description": "Refresh token của phiên cần kết thúc; bỏ trống để kết thúc mọi phiên"
					}
				}
			},
			"ForgotPasswordDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"example": "sv2026001@hcmut.edu.vn"
					}
				},
				"required": ["email"]
			},
			"ResetPasswordDto": {
				"type": "object",
				"properties": {
					"token": {
						"type": "string",
						"example": "4f2a9c1d8b3e5a7f0c6d2e4b8a1f3c5d",
						"description": "Token một lần nhận từ email (dev: ghi ra log của server)"
					},
					"password": {
						"type": "string",
						"example": "Abcd@1234"
					}
				},
				"required": ["token", "password"]
			},
			"ChangePasswordDto": {
				"type": "object",
				"properties": {
					"currentPassword": {
						"type": "string",
						"example": "Abcd@1234"
					},
					"newPassword": {
						"type": "string",
						"example": "Xyz@9876"
					}
				},
				"required": ["currentPassword", "newPassword"]
			},
...
			"UpdateCourseDto": {
				"type": "object",
				"properties": {
					"code": {
						"type": "string",
						"example": "CS102",
						"minLength": 2,
						"maxLength": 50
					},
					"title": {
						"type": "string",
						"minLength": 3,
						"maxLength": 255
					},
					"slug": {
						"type": "string",
						"maxLength": 280
					},
					"summary": {
						"type": "object",
						"maxLength": 500,
						"nullable": true
					},
					"description": {
						"type": "object",
						"maxLength": 20000,
						"nullable": true
					},
					"categoryId": {
						"type": "object",
						"nullable": true,
						"description": "UUID v4 hoặc null."
					},
					"ownerId": {
						"type": "string",
						"description": "Chỉ `admin` được đổi chủ sở hữu (UUID v4, vai trò `teacher`)."
					},
					"coverUrl": {
						"type": "object",
						"maxLength": 512,
						"nullable": true
					},
					"level": {
						"type": "string",
						"enum": ["beginner", "intermediate", "advanced"],
						"nullable": true,
						"description": "beginner: Cơ bản · intermediate: Trung cấp · advanced: Nâng cao"
					},
					"visibility": {
						"type": "string",
						"enum": ["public", "unlisted", "private"],
						"description": "public: Công khai · unlisted: Không niêm yết · private: Riêng tư"
					},
					"language": {
						"type": "string",
						"maxLength": 10
					},
					"semester": {
						"type": "object",
						"example": "2/2026-2027",
						"maxLength": 20,
						"nullable": true
					},
					"estimatedHours": {
						"type": "object",
						"minimum": 0,
						"maximum": 9999,
						"nullable": true,
						"description": "Cột là `numeric(5,1)` nên nhận số nguyên 0–9999."
					},
					"enrollmentOpen": {
						"type": "boolean"
					},
					"maxStudents": {
						"type": "object",
						"minimum": 1,
						"nullable": true
					}
				}
			},
...
			"CreateCohortDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"example": "Lớp K67-CS1",
						"minLength": 1,
						"maxLength": 150
					},
					"classCode": {
						"type": "object",
						"example": "L01",
						"maxLength": 50,
						"nullable": true
					},
					"groupCode": {
						"type": "object",
						"example": "CQ_HK261",
						"maxLength": 50,
						"nullable": true
					},
					"semester": {
						"type": "object",
						"example": "1/2026-2027",
						"maxLength": 20,
						"nullable": true
					},
					"startsOn": {
						"type": "object",
						"example": "2026-09-01",
						"nullable": true
					},
					"endsOn": {
						"type": "object",
						"example": "2027-01-15",
						"nullable": true
					}
				},
				"required": ["name"]
			},
...
			"UpdateCohortDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 150
					},
					"classCode": {
						"type": "object",
						"maxLength": 50,
						"nullable": true
					},
					"groupCode": {
						"type": "object",
						"maxLength": 50,
						"nullable": true
					},
					"semester": {
						"type": "object",
						"maxLength": 20,
						"nullable": true
					},
					"startsOn": {
						"type": "object",
						"example": "2026-09-01",
						"nullable": true
					},
					"endsOn": {
						"type": "object",
						"example": "2027-01-15",
						"nullable": true
					}
				}
			},
...
		}
	}
}